# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pajak Daerah Kota Jambi",
        "description": "API untuk aplikasi pembayaran pajak daerah (PBB-P2 & Pajak Usaha) Kota Jambi. Data Bapenda/Dukcapil masih menggunakan dummy service.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi akun baru",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "full_name",
                                    "email",
                                    "phone_number",
                                    "password",
                                    "password_confirmation",
                                    "pin_number"
                                ],
                                "properties": {
                                    "nik": {
                                        "type": "string",
                                        "example": "1671010101010001"
                                    },
                                    "full_name": {
                                        "type": "string",
                                        "example": "Ahmad Fauzi"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ahmad@example.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "pin_number": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal / NIK-email sudah dipakai"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login menggunakan NIK & password",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "password"
                                ],
                                "properties": {
                                    "nik": {
                                        "type": "string",
                                        "example": "1671010101010001"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil, return token"
                    },
                    "422": {
                        "description": "NIK/password salah atau akun terkunci"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout, hapus token aktif",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout berhasil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/nops": {
            "get": {
                "tags": [
                    "NOP (PBB-P2)"
                ],
                "summary": "List semua NOP milik user yang login beserta tagihannya",
                "operationId": "8832f8ea5866dadffb7416642cbbbcd7",
                "responses": {
                    "200": {
                        "description": "Daftar NOP berhasil diambil"
                    },
                    "401": {
                        "description": "Belum login / token tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "NOP (PBB-P2)"
                ],
                "summary": "Daftarkan NOP baru (dicek eksistensinya ke Bapenda)",
                "operationId": "69913bc64282cc8970a2ad86291b4aac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nop_number"
                                ],
                                "properties": {
                                    "nop_number": {
                                        "type": "string",
                                        "example": "3671010203040001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "NOP berhasil didaftarkan, langsung berisi tagihan"
                    },
                    "422": {
                        "description": "NOP tidak ditemukan di sistem Bapenda"
                    },
                    "429": {
                        "description": "Terlalu banyak percobaan, coba lagi nanti"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/nops/{id}": {
            "get": {
                "tags": [
                    "NOP (PBB-P2)"
                ],
                "summary": "Detail 1 NOP beserta tagihannya",
                "operationId": "46b5374cf3eb9041d1984f4a9a811e4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail NOP"
                    },
                    "404": {
                        "description": "NOP tidak ditemukan / bukan milik user ini"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/nops/{id}/refresh": {
            "post": {
                "tags": [
                    "NOP (PBB-P2)"
                ],
                "summary": "Tarik ulang tagihan terbaru dari Bapenda untuk NOP ini",
                "operationId": "02dc18bf0a199d54876c65c0d3983f87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tagihan berhasil disinkron ulang"
                    },
                    "404": {
                        "description": "NOP tidak ditemukan / bukan milik user ini"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}